refactor: move some functions to floatwin.c
