+++ src/nvim/winfloat.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "nvim/buffer_defs.h"
+
+#ifdef INCLUDE_GENERATED_DECLARATIONS
+# include "winfloat.h.generated.h"
+#endif
